refactoring: extract lanes code from scope into a new class Lanes
